requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
mysqlclient==2.2.8
pillow==12.1.1
python-dotenv==1.2.2
python-env==1.0.0
sqlparse==0.5.5
tzdata==2025.3
